è stato creato il file grafici.py per l'ottenimento dei grafici e dei file necessari per la preparazione della presentazione è stato corretto il file chart.py

+++ chart.py
@@ -2,12 +2,16 @@
 import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
+import os
 
-
+def pie_perc(pct):
+    return f'{pct:.1f}%' if pct > 2.5 else ''
+def label_filter(pct, label):
+    return label if pct > 2.5 else ''
 
 #funzione per il plot della silhouette: essa prende in ingresso tutti i valori della silhouette e li plotta
 #in un grafico a barre orizzontali
-def plot_silhouette(silhouette_vals:pd.DataFrame, cluster_label = 0):
+def plot_silhouette(silhouette_vals:pd.DataFrame, cluster_label = 0, show = True, save = False, directory:str = ''):
     # Plot dei valori di silhouette
     plt.figure(figsize=(10, 6))
     ax = plt.axes()
@@ -17,11 +21,8 @@ def plot_silhouette(silhouette_vals:pd.DataFrame, cluster_label = 0):
     size_cluster = silhouette_vals.shape[0]
     y_upper = y_lower + size_cluster
 
-    cmap = plt.cm.get_cmap('RdYlGn')
-    colors = cmap(np.linspace(0, 1, len(silhouette_vals)))
-
-    #color = 'b' # colore blu
-    plt.fill_betweenx(np.arange(y_lower, y_upper), 0, silhouette_vals, facecolor=colors, edgecolor=colors, alpha=0.7)
+    color = 'b' # colore blu
+    plt.fill_betweenx(np.arange(y_lower, y_upper), 0, silhouette_vals, facecolor=color, edgecolor=color, alpha=0.7)
 
     #plt.text(-0.05, y_lower + 0.5 * size_cluster, f'Cluster: {cluster_label}')
     y_lower = y_upper + 10  # 10 per separare i plot dei cluster
@@ -31,10 +32,18 @@ def plot_silhouette(silhouette_vals:pd.DataFrame, cluster_label = 0):
     plt.title(f"Plot dei coefficienti di silhouette per il cluster {cluster_label}")
     plt.xlabel("Valore del coefficiente di silhouette")
     plt.ylabel("Cluster")
-    plt.show()
+    if save:
+        if os.path.exists(directory):
+            plt.savefig(f'{directory}silhouette_cluster_{cluster_label}.png', dpi = 500)
+        else:
+            os.mkdir(directory)
+            plt.savefig(f'{directory}silhouette_cluster_{cluster_label}.png', dpi = 500)
+    if show:
+        plt.show()
+    plt.close()
 
 # funzione per il plot della silhouette ma stavolta le barre sono colorate in funzione del valore del coefficiente
-def plot_silhouette_colors(silhouette_vals:pd.DataFrame, cluster_label = 0):
+def plot_silhouette_colors(silhouette_vals:pd.DataFrame, cluster_label = 0, show = True, save = False, directory:str = ''):
     # Plot dei valori di silhouette
     # Numero di cluster (lunghezza dei valori della silhouette)
     n_points = len(silhouette_vals)
@@ -59,10 +68,18 @@ def plot_silhouette_colors(silhouette_vals:pd.DataFrame, cluster_label = 0):
     plt.title(f"Plot dei coefficienti di silhouette per il cluster {cluster_label}")
     plt.xlabel("Valore del coefficiente di silhouette")
     plt.ylabel("Cluster")
-    plt.show()
+    if save:
+        if os.path.exists(directory):
+            plt.savefig(f'{directory}silhouette_color_cluster_{cluster_label}.png', dpi = 500)
+        else:
+            os.mkdir(directory)
+            plt.savefig(f'{directory}silhouette_color_cluster_{cluster_label}.png', dpi = 500)
+    if show:
+        plt.show()
+    plt.close()
 
 # funzione per il calcolo dei valori presenti in una features, delle occorrenze e delle percentuali sul totale
-def feature_values(data:pd.DataFrame, feature:str):
+def feature_values(data:pd.DataFrame, feature:str, save = True, directory:str = '', cluster_label = 0):
     # Calcola le occorrenze di ciascuna categoria
     counts = data[feature].value_counts()
     # Calcola la percentuale di occorrenza
@@ -75,32 +92,64 @@ def feature_values(data:pd.DataFrame, feature:str):
     })
     
     # Salva il DataFrame in un file CSV
-    csv_feature_analysis = f"{feature}_analysis.csv"
-    result_df.to_csv(csv_feature_analysis, index=False)
+    if save:
+        if os.path.exists(directory):
+            csv_feature_analysis = f"{directory}{feature}_analysis_cluster_{cluster_label}.csv"
+            result_df.to_csv(csv_feature_analysis, index=False)
+        else:
+            os.mkdir(directory)
+            csv_feature_analysis = f"{directory}{feature}_analysis_cluster_{cluster_label}.csv"
+            result_df.to_csv(csv_feature_analysis, index=False)
 
 # funzione per il plot di un grafico a torta
-def plot_pie(data:pd.DataFrame, feature:str, titolo:str = 'Nome Feature'):
+def plot_pie(data:pd.DataFrame, feature:str, titolo:str = 'Nome Feature', show = True, save = False, directory:str = '', cluster_label = 0):
     counts = data[feature].value_counts()
     cmap = plt.cm.get_cmap('rainbow')
     # Genera una lista di colori usando linspace per ottenere esattamente il numero di colori che servono
     colors = cmap(np.linspace(0, 1, len(counts)))
 
     # Plot del grafico a torta
-    plt.figure(figsize=(6, 6))
-    plt.pie(counts, labels=counts.index, autopct='%1.1f%%', startangle=90, colors=colors)
-    plt.title(f"Distribuzione della feature {titolo}")
-    plt.show()
+    plt.figure(figsize=(13, 8))
+    font_size = max(10, min(12, 30 - len(counts)))
+
+    percentages = [count / sum(counts) * 100 for count in counts]
+
+    # Applica le etichette solo per percentuali > 2.5
+    filtered_labels = [label_filter(pct, label) for pct, label in zip(percentages, counts.index)]
+
+    plt.pie(counts, labels=filtered_labels, autopct=lambda pct:pie_perc(pct), startangle=90, colors=colors, textprops={'fontsize': font_size})
+    plt.title(f"Distribuzione della feature {titolo} nel cluster {cluster_label}")
+    if save:
+        if os.path.exists(directory):
+            plt.savefig(f'{directory}pie_plot_{feature}_cluster_{cluster_label}.png', dpi = 1000)
+        else:
+            os.mkdir(directory)
+            plt.savefig(f'{directory}pie_plot_{feature}_cluster_{cluster_label}.png', dpi = 1000)
+    if show:
+        plt.show()
+    plt.close()
 
 # funzione per il plot di un grafico a barre
-def plot_bar(data:pd.DataFrame, feature:str, titolo:str = 'Nome Feature'):
+def plot_bar(data:pd.DataFrame, feature:str, titolo:str = 'Nome Feature', cluster_label = 0, show = True, save = False, directory:str = ''):
     counts = data[feature].value_counts()
     cmap = plt.cm.get_cmap('rainbow')
     colors = cmap(np.linspace(0, 1, len(counts)))
     # Crea il grafico a barre
-    plt.figure(figsize=(8, 6))
+    plt.figure(figsize=(10, 8))
     counts.plot(kind='bar', color=colors)
-    plt.title('Occorrenze della feature ' + titolo)
+    plt.title('Occorrenze della feature ' + titolo +' nel cluster ' + str(cluster_label))
     plt.xlabel('Valore')
     plt.ylabel('Occorrenze')
-    plt.xticks(rotation=75)  # Mantiene le etichette orizzontali
-    plt.show()
+    font_size = max(5, min(10, 30 - len(counts)))
+    plt.xticks(rotation=90, fontsize = font_size)  # Mantiene le etichette orizzontali e imposta la dimensione del font
+    plt.subplots_adjust(bottom=0.35)
+    if save:
+        if os.path.exists(directory):
+            plt.savefig(f'{directory}bar_plot_{feature}_cluster{cluster_label}.png', dpi = 1000)
+        else:
+            os.mkdir(directory)
+            plt.savefig(f'{directory}bar_plot_{feature}_cluster{cluster_label}.png', dpi = 1000)
+    if show:
+        plt.show()
+    plt.close()
+
